requirements.txt
moviepy==1.0.3
Pillow<10
